refactor(protocol): move submitRouting into the package
decideSubmitRoute, decideSteerRoute and isTurnActive are pure routing
rules over SendQueueAvailability, imported by the web composer and by
mobile-native through a deep relative path into the web tree. They move
into the package beside the availability table they read (SDK plan row
C10); the test file moves with them unchanged and stays the oracle.

Relocating a module is not enough to make it reachable by package name,
so this also does the phase-C plumbing: the module joins
tsconfig.build.json's files, index.ts re-exports its three functions and
two route types, and the qualification manifest names them under the
root specifier with a real smoke call for each. Without the manifest
entry the gate passes while saying nothing about the module.

Both importers are rewritten - the web composer to the relative protocol
path its neighbours use, mobile-native to the same deep path retargeted -
and the four scenario cards that cite the file by directory are
repointed, which the source-citation audit would otherwise catch.
