+++ cmd/evener-hub/frontend/src/protocol/README.md
@@ -7,7 +7,8 @@ types, wire errors with their session classifiers, the rejection classifier
 and the user-facing message helpers every failure display goes through, the
 pure question formatter, the thread view model and its notification reducer,
 the activity tree parser, merge and disclosure rules, the job log tail parser,
-the send/queue availability table, the stable delegate status rule, and the
+the send/queue availability table, the send/steer/queue/drain routing
+decisions a composer makes off it, the stable delegate status rule, and the
 doc-pane URL builders, which hang their hrefs off a base origin the host
 supplies (empty for a same-origin web page). The doc-pane data layer is
 published at the `./docContent` subpath as well, where `readDocFile` takes the
