+++ cmd/evener-hub/frontend/src/protocol/tsconfig.build.json
@@ -29,6 +29,7 @@
     "sendQueueAvailability.ts",
     "sessionErrors.ts",
     "stableDelegate.ts",
-    "docContent.ts"
+    "docContent.ts",
+    "submitRouting.ts"
   ]
 }
