+++ mobile-native/src/composerSteering.ts
@@ -1,4 +1,4 @@
-import { decideSteerRoute } from "../../cmd/evener-hub/frontend/src/panes/session/composer/submitRouting";
+import { decideSteerRoute } from "../../cmd/evener-hub/frontend/src/protocol/submitRouting";
 import type { InputItem } from "../../cmd/evener-hub/frontend/src/protocol/types.gen";
 import type { ConversationService } from "../../mobile/src/services/conversation";
 import type { createConversationStore } from "../../mobile/src/state/conversation";
